moved categories query to const
